\name{RcppModels-package}
\alias{RcppModels-package}
\alias{RcppModels}
\docType{package}
\title{
  Reference classes for linear and generalized linear models
}
\description{
  This package provides reference classes, defined through Rcpp Modules,
  for statistical models defined by linear predictor expressions.
}
\details{
  \tabular{ll}{
Package: \tab RcppModels\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-20\cr
License: \tab GPL (>= 2)
LazyLoad: \tab yes\cr
  }
}
\author{
<Rcpp-core@r-forge.wu-wien.ac.at>
}
\keyword{ package }
%\seealso{}
%\examples{}
